# Conflicts:
#   a406_backend/src/main/java/com/social/a406/domain/board/repository/BoardImageRepository.java
#   a406_backend/src/main/java/com/social/a406/domain/board/repository/BoardRepository.java
#   a406_backend/src/main/java/com/social/a406/domain/comment/repository/CommentRepository.java
#   a406_backend/src/main/java/com/social/a406/domain/comment/service/CommentService.java
#   a406_backend/src/main/java/com/social/a406/domain/commentReply/repository/ReplyRepository.java
#   a406_backend/src/main/java/com/social/a406/domain/voiceBubble/controller/VoiceController.java

+++ a406_backend/src/main/java/com/social/a406/domain/board/dto/BoardResponse.java
@@ -18,7 +18,7 @@ public class BoardResponse {
 
     private Double x;
     private Double y;
-    private Integer z;
+    private Double z;
     private String keyword;
     private Integer pageNumber;
